[NEW] Add Odoo-Development Repository for Labiso GmbH
This commit adds the Odoo-Development repository for Labiso GmbH. The repository serves as a common development environment for Odoo, using Visual Studio Code's devcontainers. It includes an overview of the project, instructions on how to get started and use the environment, and information on its structure. Additionally, it provides guidelines on contributing to the project and support options. Finally, it is licensed under the MIT License.